// src/api/generated/model/createProductCommand.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with Keycloak Authentication and Versioning
 * OpenAPI spec version: v1
 */

export interface CreateProductCommand {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  price?: number;
  categoryId?: string;
  stockQuantity?: number;
}